sustainable-smart-city-assistant/requirements.txt
streamlit
fastapi
uvicorn
requests
python-dotenv
sentence-transformers
pydantic-settings
pinecone-client
scikit-learn
pandas
matplotlib
python-multipart
